+++ computenest-20210601/src/Models/ListServiceInstanceResourcesRequest/filters.php
@@ -1,0 +1,67 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\ComputeNest\V20210601\Models\ListServiceInstanceResourcesRequest;
+
+use AlibabaCloud\Tea\Model;
+
+class filters extends Model
+{
+    /**
+     * @description Vaild values:
+     * - ResourceARN
+     * @example ExpireTimeStart
+     *
+     * @var string
+     */
+    public $name;
+
+    /**
+     * @description The value of the filter condition.
+     *
+     * @var string[]
+     */
+    public $values;
+    protected $_name = [
+        'name'   => 'Name',
+        'values' => 'Values',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->name) {
+            $res['Name'] = $this->name;
+        }
+        if (null !== $this->values) {
+            $res['Values'] = $this->values;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return filters
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Name'])) {
+            $model->name = $map['Name'];
+        }
+        if (isset($map['Values'])) {
+            if (!empty($map['Values'])) {
+                $model->values = $map['Values'];
+            }
+        }
+
+        return $model;
+    }
+}